Moved ipc/pubsub_socket.h to ipc/socket.h and documented functions. Also removed empty integration test.
