--- ipc/test/socket_server.c
@@ -1,37 +1,0 @@
-/*
- * Copyright (C) 2017 Kubos Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#include "ipc/pubsub_socket.h"
-#include <cmocka.h>
-#include <csp/arch/csp_thread.h>
-
-#define TEST_ADDRESS 11
-#define TEST_PORT 8181
-
-static void test_server_setup(void ** arg)
-{
-    socket_conn conn;
-    assert_true(kprv_socket_server_setup(&conn, TEST_PORT, 1));
-}
-
-int main(void)
-{
-    const struct CMUnitTest tests[] = {
-        cmocka_unit_test(test_server_setup)
-    };
-
-    return cmocka_run_group_tests(tests, NULL, NULL);
-}
